backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
pynacl==1.5.0
python-dotenv==1.0.1
pydantic-settings==2.7.0
pydantic[email]==2.10.4
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.2
aiosqlite==0.20.0
